CollabRate/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.4.1
django-environ==0.12.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.2
whitenoise==6.9.0
